pyproject.toml
[project]
name = "threat-intel-vector-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "google-generativeai>=0.8.6",
    "langchain-google-genai>=4.2.1",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
    "uvicorn>=0.41.0",
]
